# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tıraş - Barber Appointment SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /barbers/{barberId}/services:
    {
      "get":
        {
          "operationId": "listServices",
          "parameters":
            [
              {
                "name": "barberId",
                "in": "path",
                "required": true,
                "schema": { "type": "integer" },
              },
            ],
          "responses":
            {
              "200":
                {
                  "description": "Success",
                  "content":
                    {
                      "application/json":
                        {
                          "schema":
                            {
                              "type": "array",
                              "items":
                                {
                                  "type": "object",
                                  "properties":
                                    {
                                      "id": { "type": "integer" },
                                      "barberId": { "type": "integer" },
                                      "name": { "type": "string" },
                                      "description": { "type": "string" },
                                      "priceKurus": { "type": "integer" },
                                      "durationMinutes": { "type": "integer" },
                                      "bufferMinutes": { "type": "integer" },
                                      "isActive": { "type": "boolean" },
                                    },
                                  "required":
                                    [
                                      "id",
                                      "barberId",
                                      "name",
                                      "description",
                                      "priceKurus",
                                      "durationMinutes",
                                      "bufferMinutes",
                                      "isActive",
                                    ],
                                },
                            },
                        },
                    },
                },
              "400": { "description": "Invalid input" },
              "401": { "description": "Unauthenticated" },
              "403": { "description": "Forbidden" },
              "409": { "description": "Conflict" },
            },
        },
    }
  /barbers/me/services:
    {
      "post":
        {
          "operationId": "createService",
          "security": [{ "bearerAuth": [] }],
          "parameters": [],
          "requestBody":
            {
              "required": true,
              "content":
                {
                  "application/json":
                    {
                      "schema":
                        {
                          "type": "object",
                          "properties":
                            {
                              "name": { "type": "string" },
                              "description": { "type": "string" },
                              "priceKurus": { "type": "integer" },
                              "durationMinutes": { "type": "integer" },
                              "bufferMinutes": { "type": "integer" },
                            },
                          "required": ["name", "priceKurus", "durationMinutes"],
                        },
                    },
                },
            },
          "responses":
            {
              "201":
                {
                  "description": "Success",
                  "content":
                    {
                      "application/json":
                        {
                          "schema":
                            {
                              "type": "object",
                              "properties":
                                {
                                  "id": { "type": "integer" },
                                  "barberId": { "type": "integer" },
                                  "name": { "type": "string" },
                                  "description": { "type": "string" },
                                  "priceKurus": { "type": "integer" },
                                  "durationMinutes": { "type": "integer" },
                                  "bufferMinutes": { "type": "integer" },
                                  "isActive": { "type": "boolean" },
                                },
                              "required":
                                [
                                  "id",
                                  "barberId",
                                  "name",
                                  "description",
                                  "priceKurus",
                                  "durationMinutes",
                                  "bufferMinutes",
                                  "isActive",
                                ],
                            },
                        },
                    },
                },
              "400": { "description": "Invalid input" },
              "401": { "description": "Unauthenticated" },
              "403": { "description": "Forbidden" },
              "409": { "description": "Conflict" },
            },
        },
    }
  /barbers/me/services/{id}:
    {
      "delete":
        {
          "operationId": "archiveService",
          "security": [{ "bearerAuth": [] }],
          "parameters":
            [
              {
                "name": "id",
                "in": "path",
                "required": true,
                "schema": { "type": "integer" },
              },
            ],
          "responses":
            {
              "204": { "description": "Success" },
              "400": { "description": "Invalid input" },
              "401": { "description": "Unauthenticated" },
              "403": { "description": "Forbidden" },
              "409": { "description": "Conflict" },
            },
        },
      "patch":
        {
          "operationId": "updateService",
          "security": [{ "bearerAuth": [] }],
          "parameters":
            [
              {
                "name": "id",
                "in": "path",
                "required": true,
                "schema": { "type": "integer" },
              },
            ],
          "requestBody":
            {
              "required": true,
              "content":
                {
                  "application/json":
                    {
                      "schema":
                        {
                          "type": "object",
                          "properties":
                            {
                              "name": { "type": "string" },
                              "description": { "type": "string" },
                              "priceKurus": { "type": "integer" },
                              "durationMinutes": { "type": "integer" },
                              "bufferMinutes": { "type": "integer" },
                            },
                          "required": ["name", "priceKurus", "durationMinutes"],
                        },
                    },
                },
            },
          "responses":
            {
              "200":
                {
                  "description": "Success",
                  "content":
                    {
                      "application/json":
                        {
                          "schema":
                            {
                              "type": "object",
                              "properties":
                                {
                                  "id": { "type": "integer" },
                                  "barberId": { "type": "integer" },
                                  "name": { "type": "string" },
                                  "description": { "type": "string" },
                                  "priceKurus": { "type": "integer" },
                                  "durationMinutes": { "type": "integer" },
                                  "bufferMinutes": { "type": "integer" },
                                  "isActive": { "type": "boolean" },
                                },
                              "required":
                                [
                                  "id",
                                  "barberId",
                                  "name",
                                  "description",
                                  "priceKurus",
                                  "durationMinutes",
                                  "bufferMinutes",
                                  "isActive",
                                ],
                            },
                        },
                    },
                },
              "400": { "description": "Invalid input" },
              "401": { "description": "Unauthenticated" },
              "403": { "description": "Forbidden" },
              "409": { "description": "Conflict" },
            },
        },
    }
  /barbers/me/access:
    {
      "get":
        {
          "operationId": "getBarberAccess",
          "security": [{ "bearerAuth": [] }],
          "parameters": [],
          "responses":
            {
              "200":
                {
                  "description": "Success",
                  "content":
                    {
                      "application/json":
                        {
                          "schema":
                            {
                              "type": "object",
                              "properties":
                                {
                                  "barberId": { "type": "integer" },
                                  "code": { "type": "string" },
                                },
                              "required": ["barberId", "code"],
                            },
                        },
                    },
                },
              "400": { "description": "Invalid input" },
              "401": { "description": "Unauthenticated" },
              "403": { "description": "Forbidden" },
              "409": { "description": "Conflict" },
            },
        },
    }
  /customers/me/barbers/join:
    {
      "post":
        {
          "operationId": "joinBarber",
          "security": [{ "bearerAuth": [] }],
          "parameters": [],
          "requestBody":
            {
              "required": true,
              "content":
                {
                  "application/json":
                    {
                      "schema":
                        {
                          "type": "object",
                          "properties": { "code": { "type": "string" } },
                          "required": ["code"],
                        },
                    },
                },
            },
          "responses":
            {
              "200":
                {
                  "description": "Success",
                  "content":
                    {
                      "application/json":
                        {
                          "schema":
                            {
                              "type": "object",
                              "properties":
                                { "barberId": { "type": "integer" } },
                              "required": ["barberId"],
                            },
                        },
                    },
                },
              "400": { "description": "Invalid input" },
              "401": { "description": "Unauthenticated" },
              "403": { "description": "Forbidden" },
              "409": { "description": "Conflict" },
            },
        },
    }
  /customers/me/barbers:
    {
      "get":
        {
          "operationId": "listMyBarbers",
          "security": [{ "bearerAuth": [] }],
          "parameters": [],
          "responses":
            {
              "200":
                {
                  "description": "Success",
                  "content":
                    {
                      "application/json":
                        {
                          "schema":
                            {
                              "type": "array",
                              "items":
                                {
                                  "type": "object",
                                  "properties":
                                    {
                                      "id": { "type": "integer" },
                                      "shopName": { "type": "string" },
                                      "shopAddress":
                                        { "type": "string", "nullable": true },
                                    },
                                  "required": ["id", "shopName"],
                                },
                            },
                        },
                    },
                },
              "400": { "description": "Invalid input" },
              "401": { "description": "Unauthenticated" },
              "403": { "description": "Forbidden" },
              "409": { "description": "Conflict" },
            },
        },
    }
  /barbers/me/customers:
    {
      "get":
        {
          "operationId": "listBarberCustomers",
          "security": [{ "bearerAuth": [] }],
          "parameters": [],
          "responses":
            {
              "200":
                {
                  "description": "Success",
                  "content":
                    {
                      "application/json":
                        {
                          "schema":
                            {
                              "type": "array",
                              "items":
                                {
                                  "type": "object",
                                  "properties":
                                    {
                                      "id": { "type": "integer" },
                                      "customerId": { "type": "integer" },
                                      "name": { "type": "string" },
                                      "phone":
                                        { "type": "string", "nullable": true },
                                      "privateNotes": { "type": "string" },
                                      "tags": { "type": "string" },
                                      "visits": { "type": "integer" },
                                      "noShows": { "type": "integer" },
                                      "lastVisit":
                                        { "type": "string", "nullable": true },
                                    },
                                  "required":
                                    [
                                      "id",
                                      "customerId",
                                      "name",
                                      "privateNotes",
                                      "tags",
                                      "visits",
                                      "noShows",
                                    ],
                                },
                            },
                        },
                    },
                },
              "400": { "description": "Invalid input" },
              "401": { "description": "Unauthenticated" },
              "403": { "description": "Forbidden" },
              "409": { "description": "Conflict" },
            },
        },
    }
  /barbers/me/customers/{id}:
    {
      "patch":
        {
          "operationId": "updateBarberCustomer",
          "security": [{ "bearerAuth": [] }],
          "parameters":
            [
              {
                "name": "id",
                "in": "path",
                "required": true,
                "schema": { "type": "integer" },
              },
            ],
          "requestBody":
            {
              "required": true,
              "content":
                {
                  "application/json":
                    {
                      "schema":
                        {
                          "type": "object",
                          "properties":
                            {
                              "privateNotes": { "type": "string" },
                              "tags": { "type": "string" },
                            },
                          "required": ["privateNotes", "tags"],
                        },
                    },
                },
            },
          "responses":
            {
              "200":
                {
                  "description": "Success",
                  "content":
                    {
                      "application/json":
                        {
                          "schema":
                            {
                              "type": "object",
                              "properties": { "id": { "type": "integer" } },
                              "required": ["id"],
                            },
                        },
                    },
                },
              "400": { "description": "Invalid input" },
              "401": { "description": "Unauthenticated" },
              "403": { "description": "Forbidden" },
              "409": { "description": "Conflict" },
            },
        },
    }
...
  /barbers/me/exceptions/{id}:
    {
      "delete":
        {
          "operationId": "deleteCalendarException",
          "security": [{ "bearerAuth": [] }],
          "parameters":
            [
              {
                "name": "id",
                "in": "path",
                "required": true,
                "schema": { "type": "integer" },
              },
            ],
          "responses":
            {
              "204": { "description": "Success" },
              "400": { "description": "Invalid input" },
              "401": { "description": "Unauthenticated" },
              "403": { "description": "Forbidden" },
              "409": { "description": "Conflict" },
            },
        },
    }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...